+++ MagicAPI/src/main/java/com/elmakers/mine/bukkit/api/magic/Mage.java
@@ -576,6 +576,8 @@ public interface Mage extends CostReducer, CooldownReducer {
     Entity getTopDamager();
     @Nullable
     Entity getLastDamageTarget();
+    @Nullable
+    Long getLastTrigger(String trigger);
 
     /**
      * This will return properties for the mage's active wand, if holding one, or if not then the active class.